Cleaned up HUD and added parallax background
Cleaned up the HUD system so that all changes to it are handled by one
script, instead of scattered around in the player scripts

Cleaned up the player health script, all damage is now controlled
centrally in that script

Switched the scrolling text to a notification text, that moves slightly
upwards then fading away, these notifications are placed in a queue, and
will not overlap. (Speed and lifetime of the notifications are changeable)

Added a parallax background script, with values that can hange both the
size and scroll speed of the background

Added error messages to the various scripts, they will specify which
components are missing for the systems to function properly. If these
components are missing, the scripts will destroy themselves, making sure
no other parts of the game is affected negatively

Added prefabs for the inventory and HUD systems, the only requirement for
these to work should now be to place them in the scene