Added a reworked prefab spawner and preview
Added a prefab spawner with the following features
-Can select which prefab to spawn
-Can select min and max scale factor
-Can select amount of prefabs to spawn
-Can select the radius where they can be spawned within

The prefab spawner as a circle that previews the radius, this circle is
visible when selectinf the prefab spawner in the inventory

Changed the player health system, when colliding with objects, the damage
recieved is scaled based on how large the collided object's mass is. The
damage now also starts at 0, instead of the damage threshold value. (A bit
more fair, no big change)