Fixed player velocity clamping and added mineables
Fixed an issue with the player velocity clamping that caused the player to
be unable to move faster than the max speed. This would be an issue if the
player were to be pushed by a fast moving object or when a boost powerup
would be used.

Added a mineable script, when triggered it spawns an item at the mining
location.

Added mining deposit prefabs, both standalone and bound to asteroids.

Changed the enemy layers and tags, the format will now be "Tag: Faction,
Layer: Enemy", instead of "Tag: Enemy, Layer: Faction" e.g. "Tag: Pirate,
Layer: Enemy"