Added level select and level objectives
Added a level select system, levels are unlocked by completing other
levels. All level data is stored in a json database and easily modified,
the actual levels themselves are stored as scenes

Added level objectives, these objectives can be placed and linked together
to form objectives that the player must complete, such as gathering a
specific item or moving to a certain area. When a level objective is
completed, objects may be activated or instantiated, for example to create
an ambush or to transition to the hub scene