Level 100 Implemented
Added the scene level_100
Added rotating lasers, with basic sprites and two related scripts.
Added doors, with 3 sprites for their different states.
Added a fusion laser which pulsates, and kills the player on impact.
Added a scenic script for spawning ships from the hangar.
Placed objective points which interact with the level, and lets the player
win in the end.
Added textures, sprites, and prefabs for the level.