Control and camera polish
The camera will now track a position ahead of the player, instead of the
player's current position. In gameplay, this makes it easier to avoid
obstacles, as the path ahead of you is visible on screen, as well as on
the minimap

Integrated the reverse with the brake, by holding down the boost button
while pressing to move backwards/brake, you will start reversing. This can
be used to get unstuck from certain geometry